backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
grpcio
grpcio-tools